+++ components/form-builder/app/branding/index.ts
@@ -1,3 +1,4 @@
 export { Logos } from "./Logos";
 export { Branding } from "./Branding";
+export { BrandingRequestForm } from "./BrandingRequestForm";
 export { options } from "./options";
