+++ pages/form-builder/settings/branding-request.tsx
@@ -1,0 +1,60 @@
+import React, { ReactElement } from "react";
+import { useTranslation } from "next-i18next";
+import { serverSideTranslations } from "next-i18next/serverSideTranslations";
+import { GetServerSideProps } from "next";
+import { getServerSession } from "next-auth";
+
+import { PageProps } from "@lib/types";
+import { getPublicTemplateByID } from "@lib/templates";
+import { authOptions } from "@pages/api/auth/[...nextauth]";
+import { NextPageWithLayout } from "../../_app";
+import { Template, PageTemplate } from "@components/form-builder/app";
+import { BrandingRequestForm } from "@components/form-builder/app/branding/";
+
+const Page: NextPageWithLayout<PageProps> = ({ publicForm }: PageProps) => {
+  const { t } = useTranslation("form-builder");
+  const title = `${t("branding.heading")} — ${t("gcForms")}`;
+  return (
+    <PageTemplate title={title}>
+      <BrandingRequestForm formRecord={publicForm} />
+    </PageTemplate>
+  );
+};
+
+Page.getLayout = (page: ReactElement) => {
+  return <Template page={page} isFormBuilder />;
+};
+
+export const getServerSideProps: GetServerSideProps = async ({
+  query: { params },
+  locale,
+  req,
+  res,
+}) => {
+  const [formID = null] = params || [];
+
+  // @TODO: get formID from somwehere
+  const publicForm = await getPublicTemplateByID("clen2l6g70194exe5kp7zl0xp");
+
+  const session = await getServerSession(req, res, authOptions);
+
+  if (session && !session.user.acceptableUse) {
+    // If they haven't agreed to Acceptable Use redirect to policy page for acceptance
+    return {
+      redirect: {
+        destination: `/${locale}/auth/policy`,
+        permanent: false,
+      },
+    };
+  }
+
+  return {
+    props: {
+      publicForm,
+      ...(locale &&
+        (await serverSideTranslations(locale, ["common", "form-builder"], null, ["fr", "en"]))),
+    },
+  };
+};
+
+export default Page;
